+++ polishing/subs_matrix.h
@@ -14,14 +14,13 @@ class SubstitutionMatrix
 {
 public:
 	SubstitutionMatrix(const std::string& path);
-	double getScore(char v, char w) const;
-	//void printMatrix() const;
+	float getScore(char v, char w) const;
 private:
 	void loadMatrix(const std::string& path);
 
 	const int X_SIZE = 5;
 	const int Y_SIZE = 5;
-	std::vector<std::vector<double>> _matrix;
+	std::vector<std::vector<float>> _matrix;
 };
 
 class HopoMatrix
@@ -59,9 +58,9 @@ public:
 	typedef std::vector<Observation> ObsVector;
 
 	HopoMatrix(const std::string& fileName);
-	double getObsProb(State state, Observation observ) const
+	float getObsProb(State state, Observation observ) const
 		{return _observationProbs[state.id][observ.id];}
-	double getGenomeProb(State state) const
+	float getGenomeProb(State state) const
 		{return _genomeProbs[state.id];}
 	ObsVector knownObservations(State state) const;
 	static Observation strToObs(char mainNucl, const std::string& dnaStr, 
@@ -72,6 +71,6 @@ public:
 private:
 	void loadMatrix(const std::string& filaName);
 
-	std::vector<std::vector<double>> _observationProbs;
-	std::vector<double> 			 _genomeProbs;
+	std::vector<std::vector<float>> _observationProbs;
+	std::vector<float> 			 	_genomeProbs;
 };
